+++ src/NeuralNet/Optimizers/Adam.php
@@ -51,6 +51,20 @@ class Adam implements Optimizer, Adaptive
     protected float $normDecay;
 
     /**
+     * The opposite of the momentum decay.
+     *
+     * @var float
+     */
+    protected float $beta1;
+
+    /**
+     * The opposite of the norm decay.
+     *
+     * @var float
+     */
+    protected float $beta2;
+
+    /**
      * The parameter cache of running velocity and squared gradients.
      *
      * @var array[]
@@ -58,6 +72,13 @@ class Adam implements Optimizer, Adaptive
     protected array $cache = [
         //
     ];
+
+    /**
+     * The number of steps taken since initialization.
+     *
+     * @var int
+     */
+    protected int $t = 0;
 
     /**
      * @param float $rate
@@ -85,6 +106,8 @@ class Adam implements Optimizer, Adaptive
         $this->rate = $rate;
         $this->momentumDecay = $momentumDecay;
         $this->normDecay = $normDecay;
+        $this->beta1 = 1.0 - $momentumDecay;
+        $this->beta2 = 1.0 - $normDecay;
     }
 
     /**
@@ -116,10 +139,10 @@ class Adam implements Optimizer, Adaptive
     {
         [$velocity, $norm] = $this->cache[$param->id()];
 
-        $velocity = $velocity->multiply(1.0 - $this->momentumDecay)
+        $velocity = $velocity->multiply($this->beta1)
             ->add($gradient->multiply($this->momentumDecay));
 
-        $norm = $norm->multiply(1.0 - $this->normDecay)
+        $norm = $norm->multiply($this->beta2)
             ->add($gradient->square()->multiply($this->normDecay));
 
         $this->cache[$param->id()] = [$velocity, $norm];