+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 1.0.1
     - Prevent certain specification false negatives
     - Add extension minimum version specification
+    - Fix AdaMax optimizer when tensor extension loaded
 
 - 1.0.0
     - No changes